% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qreml_functions.R
\name{AIC.qremlModel}
\alias{AIC.qremlModel}
\title{Akaike's Information Criterion for models returned by \code{\link{qreml}}}
\usage{
\method{AIC}{qremlModel}(object, k = 2)
}
\arguments{
\item{object}{a fitted model object returned by \code{\link{qreml}}}

\item{k}{numeric, the penalty per parameter to be used; the default \code{k = 2} is the classical AIC.}
}
\value{
numeric value with the corresponding AIC (or BIC, or ..., depending on k)
}
\description{
Akaike's Information Criterion for models returned by \code{\link{qreml}}
}
\examples{
## no examples
}
