% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dgamma2}
\alias{dgamma2}
\title{Reparametrized gamma distribution}
\usage{
dgamma2(x, mu, sigma)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu}{mean parameter, must be positive scalar.}

\item{sigma}{standard deviation parameter, must be positive scalar.}
}
\value{
Returns the gamma density function evaluated at x.
}
\description{
Returns the gamma density function reparametrized in terms of mean and standard deviation. 
This implementation allows for automatic differentiation with RTMB.
}
\examples{
dgamma2(1, 1, 1)
}
