% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{dirichlet}
\alias{dirichlet}
\alias{ddirichlet}
\title{Dirichlet distribution}
\usage{
ddirichlet(x, alpha, log = TRUE)
}
\arguments{
\item{x}{vector or matrix of quantiles}

\item{alpha}{vector or matrix of shape parameters}

\item{log}{logical; if \code{TRUE}, densities \eqn{p} are returned as \eqn{\log(p)}.}
}
\value{
\code{ddirichlet} gives the density.
}
\description{
Density of the Dirichlet distribution.
}
\details{
This implementation of \code{ddirichlet} allows for automatic differentiation with \code{RTMB}.
}
\examples{
ddirichlet(c(0.2, 0.3, 0.5), c(1, 2, 3))
}
