% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvm.R
\name{dvm}
\alias{dvm}
\title{von Mises Density Function}
\usage{
dvm(x, mu, kappa)
}
\arguments{
\item{x}{vector of angles measured in radians at which to evaluate the density function.}

\item{mu}{mean direction of the distribution measured in radians.}

\item{kappa}{non-negative numeric value for the concentration parameter of the distribution.}
}
\value{
Returns the von Mises density function evaluated at theta.
}
\description{
Returns the von Mises density function evaluated at a particular value. 
This implementation allows for automatic differentiation with RTMB.
}
\examples{
x = c(0, pi/2, pi)
dvm(x, 0, 1)
}
