% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elephant}
\alias{elephant}
\title{Elephant data set}
\format{
A data frame with 10.000 rows and 4 variables:
\describe{
  \item{tod}{time of day variable ranging from 1 to 24}
  \item{step}{hourly step lengths in kilometres}
  \item{angle}{hourly turning angles in radians}
  \item{state}{hidden state variable}
}
}
\source{
Generated for example purposes.
}
\usage{
elephant
}
\description{
Synthetic data set of hourly step lengths and turning angles of an elephant.
}
\keyword{datasets}
