% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_flies.R
\name{forward_flies}
\alias{forward_flies}
\title{Forward algorithm for fruit fly application}
\usage{
forward_flies(delta, allprobs, Gamma, DD, tod)
}
\arguments{
\item{delta}{Initial or periodically stationary distribution of length N}

\item{allprobs}{Matrix of state-dependent probabilities/ density values of dimension c(n, N)}

\item{Gamma}{Array of transition probability matrices of dimension c(N,N,L,2).}

\item{DD}{Index where DD condition starts}

\item{tod}{time of day variable}
}
\value{
Log-likelihood for given data and parameters
}
\description{
Forward algorithm for fruit fly application
}
