% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_h.R
\name{forward_h}
\alias{forward_h}
\title{R-Wrapper function for forward algorithm with homogeneous t.p.m.}
\usage{
forward_h(delta, allprobs, Gamma)
}
\arguments{
\item{delta}{initial distribution}

\item{allprobs}{allprobs matrix}

\item{Gamma}{Gamma matrix}
}
\value{
log likelihood
}
\description{
R-Wrapper function for forward algorithm with homogeneous t.p.m.
}
