% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{gdeterminant}
\alias{gdeterminant}
\title{Computes generalised determinant}
\usage{
gdeterminant(x, eps = NULL, log = TRUE)
}
\arguments{
\item{x}{symmetric matrix}

\item{eps}{eigenvalues smaller than this will be treated as zero}

\item{log}{logical. If \code{TRUE}, the log-determinant is returned. If \code{FALSE}, the determinant is returned.}
}
\value{
generalised log-determinant of \code{x}
}
\description{
Computes generalised determinant
}
