\name{Lcpp-package}
\alias{Lcpp-package}
\alias{Lcpp}
\docType{package}
\title{
\packageTitle{Lcpp}
}
\description{
\packageDescription{Lcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lcpp}
\packageIndices{Lcpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Lcpp}

Maintainer: \packageMaintainer{Lcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
