% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{logLik.LaMaModel}
\alias{logLik.LaMaModel}
\title{Extract log-likelihood from LaMaModel object}
\usage{
\method{logLik}{LaMaModel}(object, ...)
}
\arguments{
\item{object}{A model fitted using RTMB and obtained via \code{report(obj)} of class "LaMaModel"}

\item{...}{Additional arguments (not used)}
}
\value{
An object of class "logLik"
}
\description{
Extract log-likelihood from LaMaModel object
}
