% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllk1.R
\name{mllk_random}
\alias{mllk_random}
\title{Negative log Likelihood for the fly data RE model}
\usage{
mllk_random(theta.star, X, L = 48, K, M, low, up)
}
\arguments{
\item{theta.star}{unconstraint working parameter vector}

\item{X}{data}

\item{L}{number of time points in a cycle}

\item{K}{degree of trigonometric link}

\item{M}{numer of intervals for numerical approximation of the integral}

\item{low}{lower integral boundaries}

\item{up}{upper integral boundaries}
}
\value{
returns the negative log likelihood of the data given the paramter vector
}
\description{
Negative log Likelihood for the fly data RE model
}
