% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nessi}
\alias{nessi}
\title{Loch Ness Monster Acceleration Data}
\format{
A data frame with 5.000 rows and 3 variables:
\describe{
  \item{ODBA}{overall dynamci body acceleration}
  \item{logODBA}{logarithm of overall dynamic body acceleration}
  \item{state}{hidden state variable}
}
}
\source{
Generated for example purposes.
}
\usage{
nessi
}
\description{
A small group of researchers managed to put an accelerometer on the Loch Ness Monster and collected data for a few days. 
Now we have a data set of the overall dynamic body acceleration (ODBA) of the creature.
}
\keyword{datasets}
