% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pql_functions.R
\name{pql}
\alias{pql}
\title{Penalized quasi-likelihood (PQL) algorithm for mixed models}
\usage{
pql(
  pnll,
  par,
  dat,
  random,
  alpha_sm = 0.98,
  maxiter = 50,
  tol = 0.01,
  silent = TRUE,
  saveall = FALSE
)
}
\arguments{
\item{pnll}{penalized negative log-likelihood function that is structured as dictated by \code{RTMB} and uses the \code{penalty} function from \code{LaMa} to compute the penalty.}

\item{par}{Named list of initial parameters. The random effects can be vectors or matrices, the latter summarising several random effects of the same structure, each one being a row in the matrix.}

\item{dat}{Initial data list, that contains the data used in the likelihood function, hyperparameters, and the initial penalty strength that needs to be called \code{lambda} and structured as detailed in the documentation of \code{penalty}.}

\item{random}{vector of names of the random effects in \code{par} that are penalized.}

\item{alpha_sm}{optional hyperparamater for exponential smoothing of the penalty strengths. For smaller values smoother convergence is to be expected but the algorithm may need more iterations.}

\item{maxiter}{maximum number of iterations.}

\item{tol}{convergence tolerance for the penalty strength parameters.}

\item{silent}{logical, if TRUE, then the inner optimization detailes are not printed.}

\item{saveall}{logical, if TRUE, then all model objects from each iteration are saved in the final model object. Defaults to FALSE.}
}
\value{
Returns a model list influenced by the users report statements in \code{pnll}
}
\description{
This algorithm can be used very flexible to fit any kind of statistical model that involves splines or simple i.i.d. random effects with \code{RTMB} with penalties of the form\cr\cr
\eqn{0.5 \sum_{i} \lambda_i b_i^T S_i b_i}\cr\cr
PQL is typically much faster than the full Laplace approximation method, but may be slightly less accurate regarding the estimation of the penalty strength parameters.
The user has to specify the penalized negative log-likelihood function \code{pnll} structured as dictated by \code{RTMB} and use the \code{penalty} function contained in \code{LaMa} to compute the penalty inside the likelihood.
}
\examples{
# currently no example
}
