% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{pred_matrix_dens}
\alias{pred_matrix_dens}
\title{Build the prediction design matrix for a new data vector based on model_matrices object created by \code{make_matrices_dens()}}
\usage{
pred_matrix_dens(model_matrices, xnew)
}
\arguments{
\item{model_matrices}{model_matrices object as returned from \code{make_matrices_dens()}}

\item{xnew}{new data vector for which to evaluate the basis}
}
\value{
prediction design matrix for \code{xnew} with the same basis as used for \code{model_matrices}
}
\description{
Build the prediction design matrix for a new data vector based on model_matrices object created by \code{make_matrices_dens()}
}
\examples{
modmat = make_matrices_dens(x = 1:100, K = 20)
pred_matrix_dens(modmat, xnew = 1:100 - 0.5)
}
