% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix_functions.R
\name{print.LaMa_matrices}
\alias{print.LaMa_matrices}
\title{Print a \code{LaMa_matrices} object}
\usage{
\method{print}{LaMa_matrices}(x, ...)
}
\arguments{
\item{x}{object of class \code{LaMa_matrices} as returned by \code{\link{make_matrices}}}

\item{...}{ignored}
}
\value{
\code{x}, invisibly. Called for its side effect of printing.
}
\description{
Print a \code{LaMa_matrices} object
}
\seealso{
\code{\link{make_matrices}}
}
