% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pql_functions.R
\name{safeAD}
\alias{safeAD}
\title{Overload some settings to make automatic differentiation work smoothly}
\usage{
safeAD()
}
\value{
Returns nothing, but overloads some settings to make AD work.
}
\description{
Run this function in the first line of your likelihood function when using \code{RTMB} to be safe.
}
\examples{
nll = function(par){
 safeAD()
 # ...
 print("AD is safe now")
}
par = list(a = 1, b = 2)
nll(par)
}
