% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shark}
\alias{shark}
\title{Shark Acceleration Data}
\format{
A data frame with 5.000 rows and 3 variables:
\describe{
  \item{ODBA}{overall dynamci body acceleration}
  \item{logODBA}{logarithm of overall dynamic body acceleration}
  \item{state}{hidden state variable}
}
}
\source{
Generated for example purposes.
}
\usage{
shark
}
\description{
Synthetic data set of minutely overall dynamic body acceleration (ODBA) of a shark.
}
\keyword{datasets}
