% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{vm}
\alias{vm}
\alias{dvm}
\alias{pvm}
\alias{rvm}
\title{von Mises distribution}
\usage{
dvm(x, mu, kappa, log = FALSE)

pvm(q, mu, kappa, acc = 1e-20)

rvm(n, mu, kappa, wrap = TRUE)
}
\arguments{
\item{x, q}{vector of angles measured in radians at which to evaluate the density function.}

\item{mu}{mean direction of the distribution measured in radians.}

\item{kappa}{non-negative numeric value for the concentration parameter of the distribution.}

\item{log}{logical; if \code{TRUE}, densities are returned on the log scale.}

\item{acc}{accuracy of the distribution function approximation (for details see the \code{CircStats} documentation).}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{wrap}{logical; if \code{TRUE}, generated angles are wrapped to the interval [-pi, pi].}
}
\value{
\code{dvm} gives the density, \code{pvm} gives the distribution function, and \code{rvm} generates random deviates.
}
\description{
Density, distribution function and random generation for the von Mises distribution.
}
\details{
The implementation of \code{dvm} allows for automatic differentiation with \code{RTMB}. The other functions are imported from \code{CircStats}.
}
\examples{
x = rvm(2, 0, 2)
d = dvm(x, 0, 2)
p = pvm(x, 0, 2)
}
