% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dvm}
\alias{dvm}
\title{von Mises Density Function}
\usage{
dvm(x, mu, kappa, log = FALSE)
}
\arguments{
\item{x}{vector of angles measured in radians at which to evaluate the density function.}

\item{mu}{mean direction of the distribution measured in radians.}

\item{kappa}{non-negative numeric value for the concentration parameter of the distribution.}

\item{log}{logical; if TRUE, densities are returned on the log scale.}
}
\value{
Returns the density function of the von Mises density distribution evaluated at x.
}
\description{
Returns the density function of the van Mises distribution evaluated at a particular value.\cr\cr
This implementation allows for automatic differentiation with RTMB.
}
\examples{
x = c(0, pi/2, pi)
dvm(x, 0, 1)
}
