% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{Forward algorithm with homogeneous transition probability matrix}
\usage{
forward(delta, allprobs, Gamma)
}
\arguments{
\item{delta}{initial or stationary distribution}

\item{allprobs}{allprobs matrix (of dimension c(n, N))}

\item{Gamma}{Gamma matrix (of dimension c(N,N))}
}
\value{
log likelihood
}
\description{
Forward algorithm with homogeneous transition probability matrix
}
