% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{R-Wrapper function for forward algorithm in CPP for periodic variation}
\usage{
forward(delta, allprobs, Gamma, DD, tod)
}
\arguments{
\item{delta}{Initial distribution}

\item{allprobs}{allprobs matrix}

\item{tod}{time of day variable}

\item{Gamma1}{Gamma array in LD condition (dim = c(N, N, L))}

\item{Gamma2}{Gamma array in DD condition (dim = c(N, N, L))}

\item{startDD}{startindex of DD condition}
}
\value{
log likelihood
}
\description{
R-Wrapper function for forward algorithm in CPP for periodic variation
}
