% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{R-Wrapper function for forward algorithm in CPP for periodic variation}
\usage{
forward(delta, allprobs, Gamma1, Gamma2, startDD, X)
}
\arguments{
\item{delta}{Initial distribution}

\item{allprobs}{allprobs matrix}

\item{Gamma1}{Gamma matrix in LD condition (array)}

\item{Gamma2}{Gamma matrix in DD condition (array)}

\item{startDD}{startindex of DD condition}

\item{X}{data}
}
\description{
R-Wrapper function for forward algorithm in CPP for periodic variation
}
