% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_flies.R
\name{forward_flies}
\alias{forward_flies}
\title{Forward algorithm for fruit fly application}
\usage{
forward_flies(delta, allprobs, Gamma, DD, tod)
}
\arguments{
\item{delta}{initial distribution}

\item{allprobs}{allprobs matrix (of dimension c(n, N))}

\item{tod}{time of day variable}

\item{Gamma1}{Gamma array in LD condition (dim = c(N, N, L))}

\item{Gamma2}{Gamma array in DD condition (dim = c(N, N, L))}

\item{startDD}{startindex of DD condition}
}
\value{
Log-likelihood for given data and parameters
}
\description{
Forward algorithm for fruit fly application
}
