% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_g.R
\name{forward_g}
\alias{forward_g}
\title{General forward algorithm with time-varying transition probability matrix}
\usage{
forward_g(delta, Gamma, allprobs)
}
\arguments{
\item{delta}{Initial or periodically stationary distribution (of length N)}

\item{Gamma}{Pre-calculated array of Gamma matrices (of dimension c(N,N,n))}

\item{allprobs}{allprobs matrix (of dimension c(n, N))}
}
\value{
Log-likelihood for given data and parameters
}
\description{
General forward algorithm with time-varying transition probability matrix
}
