% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_g.R
\name{forward_g}
\alias{forward_g}
\title{General forward algorithm with time-varying transition probability matrix}
\usage{
forward_g(delta, allprobs, Gamma)
}
\arguments{
\item{delta}{initial or periodically stationary distribution}

\item{allprobs}{allprobs matrix (of dimension c(n, N))}

\item{Gamma}{array of Gamma matrices (of dimension c(N,N,n))}
}
\value{
log likelihood
}
\description{
General forward algorithm with time-varying transition probability matrix
}
