% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_g.R
\name{forward_g}
\alias{forward_g}
\title{R-Wrapper function for general forward algorithm with t.p.m. as an array of dim c(N,N,T)}
\usage{
forward_g(delta, allprobs, Gamma)
}
\arguments{
\item{delta}{initial distribution}

\item{allprobs}{allprobs matrix}

\item{Gamma}{array of Gamma matrices of dim c(N,N,T)}
}
\value{
log likelihood
}
\description{
R-Wrapper function for general forward algorithm with t.p.m. as an array of dim c(N,N,T)
}
