% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_functions.R
\name{forward_hsmm}
\alias{forward_hsmm}
\title{\href{https://www.taylorfrancis.com/books/mono/10.1201/b20790/hidden-markov-models-time-series-walter-zucchini-iain-macdonald-roland-langrock}{Forward algorithm} for homogeneous hidden semi-Markov models}
\usage{
forward_hsmm(
  dm,
  omega,
  allprobs,
  trackID = NULL,
  delta = NULL,
  eps = 1e-10,
  report = TRUE
)
}
\arguments{
\item{dm}{List of length N containing vectors of dwell-time probability mass functions (PMFs) for each state. The vector lengths correspond to the approximating state aggregate sizes, hence there should be little probablity mass not covered by these.}

\item{omega}{Matrix of dimension c(N,N) of conditional transition probabilites. This contains the transition probabilities given the current state is left. Hence, the diagonal elements need to be zero and the rows need to sum to one.}

\item{allprobs}{Matrix of state-dependent probabilities/ density values of dimension c(n, N)}

\item{trackID}{Optional vector of length n containing k unique IDs. If provided, the total log-likelihood will be the sum of each track's likelihood contribution.
In this case, \code{dm} can be a nested list, where the top layer contains k \code{dm} lists as described above. \code{omega} can then also be an array of dimension c(N,N,k) with one conditional transition probability matrix for each track.
Furthermore, instead of a single vector \code{delta} corresponding to the initial distribution, a delta matrix of initial distributions, of dimension c(k,N), can be provided, such that each track starts with it's own initial distribution.}

\item{delta}{Optional vector of initial state probabilities of length N. By default, the stationary distribution is computed (which is typically recommended).}

\item{eps}{Small value to avoid numerical issues in the approximating transition matrix construction.}

\item{report}{Logical, indicating whether initial distribution, approximating t.p.m. and allprobs matrix should be reported from the fitted model. Defaults to TRUE.}
}
\value{
Log-likelihood for given data and parameters
}
\description{
Hidden semi-Markov models (HSMMs) are a flexible extension of HMMs, where the state duration distribution is explicitly modelled by a distribution on the positive integers.
For direct numerical maximum likelhood estimation, HSMMs can be represented as HMMs on an enlarged state space (of size \eqn{M}) and with structured transition probabilities.

This function is designed to be used with automatic differentiation based on the \code{R} package \code{RTMB}. It will be very slow without it.
}
\examples{
# currently no examples
}
