% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{LaMaColors}
\alias{LaMaColors}
\title{Generate a colour-blind-friendly palette}
\usage{
LaMaColors(ncolors)
}
\arguments{
\item{ncolors}{Number of colours to return; at most 10}
}
\value{
A colour palette of length \code{ncolors}
}
\description{
Generate a colour-blind-friendly palette
}
\examples{
cb <- LaMaColors(3)
}
