% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{LaMaColors}
\alias{LaMaColors}
\title{Generate colour palette}
\usage{
LaMaColors(ncolors)
}
\arguments{
\item{ncolors}{Number of colours to return; at most}
}
\value{
A colour palette of 7 colors
}
\description{
Generate colour palette
}
\examples{
cb <- LaMaColors(3)
}
