% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lcpp-package.R
\docType{package}
\name{Lcpp-package}
\alias{Lcpp}
\alias{Lcpp-package}
\title{Lcpp: Faster numerical maximum likelihood estimation for HMMs and SMMs}
\description{
Forward algorithm in C++ to speed up estimation of HMMs and SSMs.
}
\author{
\strong{Maintainer}: Jan-Ole Koslik \email{jan-ole.koslik@uni-bielefeld.de}

}
\keyword{internal}
