% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lcpp-package.R
\docType{package}
\name{Lcpp-package}
\alias{Lcpp}
\alias{Lcpp-package}
\title{Lcpp: Faster numerical maximum likelihood estimation of HMMs, HSMMs and SMMs}
\description{
This package provides convenient R wrapper functions for the forward algorithm used to fit hidden Markov models (HMMs), hidden semi-Markov models (HSMMs) and state space models (SSMs) via direct numerical maximum likelihood estimation. The algorithm calculates the log-likelihood recursively as a matrix product and uses a scaling strategy to avoid numerical underflow. Implementation in C++ offers 10-20 times faster evaluation times, thus substantially speeding up estimation by numerical optimizers.
}
\author{
\strong{Maintainer}: Jan-Ole Koslik \email{jan-ole.koslik@uni-bielefeld.de}

}
\keyword{internal}
