% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qreml_functions.R
\name{logLik.qremlModel}
\alias{logLik.qremlModel}
\title{Extract log-likelihood from qremlModel object}
\usage{
\method{logLik}{qremlModel}(object, ...)
}
\arguments{
\item{object}{A fitted model of class "qremlModel"}

\item{...}{Additional arguments (not used)}
}
\value{
An object of class "logLik"
}
\description{
Extract log-likelihood from qremlModel object
}
