% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{make_matrices_dens}
\alias{make_matrices_dens}
\title{Build a standardized P-Spline design matrix and the associated P-Spline penalty matrix}
\usage{
make_matrices_dens(x, K, degree = 3, npoints = 10000, diff_order = 2)
}
\arguments{
\item{x}{data vector}

\item{K}{number of basis functions}

\item{degree}{degree of the B-spline basis functions, defaults to cubic B-Splines}

\item{npoints}{number of points used in the numerical integration for normalizing the B-spline basis functions}

\item{diff_order}{order of the difference used for the penalty matrix. Defaults to second-order differences.}
}
\value{
a list containing the design matrix Z, the penalty matrix S, the knots, the normalization weights w, the degree of the B-spline basis functions and the basis positions
}
\description{
This function builds the B-spline design matrix for a given data vector x. 
The B-spline basis functions are normalized such that the integral of each basis function is 1, hence this basis can be used for spline-based density estimation.
}
\examples{
modmat = make_matrices_dens(x = 1:100, K = 20)
}
