% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcreport.R
\name{MCreport}
\alias{MCreport}
\title{Sample parameters from approximate Gaussian posterior distribution}
\usage{
MCreport(obj, nSamples = 1000, include_random_pars = TRUE, report = FALSE, ...)
}
\arguments{
\item{obj}{Optimised \code{RTMB} object}

\item{nSamples}{Number of samples to draw}

\item{include_random_pars}{Logical; Should random parameters be included in the output?}

\item{report}{Logical; Should reported quantities be sampled as well? 
Defaults to \code{FALSE} because this may be slow depending on your model.}

\item{...}{For internal use only}
}
\value{
A list of parameter samples, each structured like the initial parameter list from \link[RTMB]{MakeADFun}
}
\description{
Efficient Monte Carlo sampling of parameters (and \code{REPORT}ed quantities) from the approximate posterior of an \code{RTMB} model.
See \link[TMB]{sdreport} for details on posterior variance-covariance in random effects models.
}
\examples{
step <- trex$step[1:1000] # subsetting trex data
N <- 2                    # 2 states

# custom likelihood
nll <- function(par) {
  getAll(par)
  Gamma <- tpm(eta)
  delta <- stationary(Gamma)
  mu <- exp(log_mu); REPORT(mu)
  sigma <- exp(log_sigma); REPORT(sigma)
  allprobs <- matrix(1, length(step), N)
  for(j in 1:N) allprobs[,j] <- dgamma2(step, mu[j], sigma[j])
  -forward(delta, Gamma, allprobs)
}

# initial parameters in named list
par0 <- list(eta = rep(-2,2), 
             log_mu = log(c(0.3, 1)), 
             log_sigma = log(c(0.2, 0.7)))
    
# constructing AD object        
obj <- MakeADFun(nll, par0, silent = TRUE)

# optimising
opt <- nlminb(obj$par, obj$fn, obj$gr)

# sampling from distribution of the MLE
samples <- MCreport(obj, nSamples = 10, report = TRUE)
}
