% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pql_functions.R
\name{pql}
\alias{pql}
\title{Penalized quasi-likelihood (PQL) algorithm for models with simple random effects}
\usage{
pql(
  pnll,
  par,
  dat,
  random,
  alpha = 0,
  maxiter = 50,
  tol = 1e-05,
  inner_tol = 1e-10,
  silent = 1,
  saveall = FALSE
)
}
\arguments{
\item{pnll}{Penalized negative log-likelihood function that is structured as dictated by \code{RTMB} and uses the \code{penalty} function from \code{LaMa} to compute the penalty.}

\item{par}{Named list of initial parameters. The random effects can be vectors or matrices, the latter summarising several random effects of the same structure, each one being a row in the matrix.}

\item{dat}{Initial data list, that contains the data used in the likelihood function, hyperparameters, and the initial penalty strength. Currently needs to be called \code{dat}. The initial penalty strength vector needs to be called \code{lambda} and has to be of length corresponding to the total number of random effects.}

\item{random}{Vector of names of the random effects in \code{par} that are penalized.}

\item{alpha}{Optional hyperparamater for exponential smoothing of the penalty strengths. For larger values smoother convergence is to be expected but the algorithm may need more iterations.}

\item{maxiter}{Maximum number of iterations.}

\item{tol}{Convergence tolerance for the penalty strength parameters.}

\item{inner_tol}{Convergence tolerance for the inner optimization.}

\item{silent}{Integer silencing level: 0 corresponds to full printing of inner and outer iteratinos, 1 to printing of outer iterations only, and 2 to no printing.}

\item{saveall}{Logical, if TRUE, then all model objects from each iteration are saved in the final model object. Defaults to FALSE.}
}
\value{
Returns a model list influenced by the users report statements in \code{pnll}
}
\description{
This algorithm can be used very flexible to fit statistical models that involves \strong{penalized splines} or simple \strong{i.i.d. random effects} with \code{RTMB} that have penalties of the form
\deqn{0.5 \sum_{i} \lambda_i b_i^T S_i b_i}
PQL is typically much faster than the full Laplace approximation method, but may be slightly less accurate regarding the estimation of the penalty strength parameters.
The user has to specify the penalized negative log-likelihood function \code{pnll} structured as dictated by \code{RTMB} and use the \code{penalty} function contained in \code{LaMa} to compute the penalty inside the likelihood.
}
\examples{
# currently no example
}
