% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_functions.R
\name{pseudo_res}
\alias{pseudo_res}
\title{Calculate pseudo-residuals}
\usage{
pseudo_res(
  obs,
  stateprobs,
  dist,
  par,
  normal = TRUE,
  discrete = NULL,
  randomise = TRUE,
  seed = NULL
)
}
\arguments{
\item{obs}{vector of continuous-valued observations (of length n)}

\item{stateprobs}{matrix of local state probabilities for each observation (of dimension c(n,N), where N is the number of states) as computed by \code{\link{stateprobs}}, \code{\link{stateprobs_g}} or \code{\link{stateprobs_p}}}

\item{dist}{character string specifying which parametric CDF to use (e.g., "norm" for normal)}

\item{par}{named parameter list for the parametric CDF (e.g. list(mean = c(1,2), sd = c(1,1)) for a normal distribution and 2 states)}

\item{normal}{logical, if \code{TRUE}, return Gaussian pseudo residuals

These will be approximately standard normally distributed if the model is correct.}

\item{discrete}{logical, if \code{TRUE}, computes discrete pseudo residuals (which slightly differ in their definition)

By default, will be determined using \code{dist} argument, but only works for standard discrete distributions.
When used with a special discrete distribution, set to \code{TRUE} manually. See \code{\link{pseudo_res_discrete}} for details.}

\item{randomise}{for discrete pseudo residuals only. Logical, if \code{TRUE}, return randomised pseudo residuals. Recommended for discrete observations.}

\item{seed}{for discrete pseudo residuals only. Integer, seed for random number generation}
}
\value{
vector of pseudo residuals
}
\description{
For HMMs, pseudo-residuals are used to assess the goodness-of-fit of the model. 
These are based on the cumulative distribution function (CDF)
\deqn{F_{X_t}(x_t) = f(x_t \mid x_1, \dots, x_{t-1}, x_{t+1}, \dots, x_T)}
and can be used to assess whether an observation is extreme relative to its model-implied distribution.

This function calculates such residuals via probability integral transform, based on the local state probabilities obtained by \code{\link{stateprobs}} or \code{\link{stateprobs_g}} and the respective parametric family.
}
\details{
When used for discrete pseudo-residuals, this function is just a wrapper for \code{\link{pseudo_res_discrete}}.
}
\examples{
## continuous-valued observations
obs = rnorm(100)
stateprobs = matrix(0.5, nrow = 100, ncol = 2)
par = list(mean = c(1,2), sd = c(1,1))
pres = pseudo_res(obs, stateprobs, "norm", par)

## discrete-valued observations
obs = rpois(100, lambda = 1)
stateprobs = matrix(0.5, nrow = 100, ncol = 2)
par = list(lambda = c(1,2))
pres = pseudo_res(obs, stateprobs, "pois", par)
}
