% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{report}
\alias{report}
\title{Get reported quantities from and \code{RTMB} object and return a LaMaModel}
\usage{
report(obj)
}
\arguments{
\item{obj}{Object returned by \code{RTMB::MakeADFun()}}
}
\value{
A model object of class "\code{LaMaModel}" containing the reported quantities from the \code{RTMB} object, along with log-likelihood, number of parameters, and number of observations.
}
\description{
Get reported quantities from and \code{RTMB} object and return a LaMaModel
}
\examples{
# no example yet
}
