% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdreportMC.R
\name{sdreportMC}
\alias{sdreportMC}
\title{Monte Carlo version of sdreport}
\usage{
sdreportMC(
  obj,
  what,
  Hessian = NULL,
  CI = FALSE,
  n = 1000,
  probs = c(0.025, 0.975)
)
}
\arguments{
\item{obj}{Object returned by \code{MakeADFun()} after optimization}

\item{what}{Vector of strings with names of parameters and \code{REPORT()}ed quantities to be reported}

\item{Hessian}{Optional Hessian matrix. If not provided, it will be computed from the object}

\item{CI}{Logical. If \code{TRUE}, only confidence intervals instead of samples will be returned}

\item{n}{Number of samples to draw from the multivariate normal distribution of the MLE}

\item{probs}{Vector of probabilities for the confidence intervals (ignored if no CIs are computed)}
}
\value{
Named list corresponding to the elements of \code{what}. Each element has the structure of the corresponding quantity with an additional dimension added for the samples.
For example, if a quantity is a vector, the list contains a matrix. If a quantity is a matrix, the list contains an array.
}
\description{
After optimization of an AD model, sdreportMC can be used to calculate samples of confidence intervals of all model parameters and \code{REPORT()}ed quantities
including non linear functions of random effects and parameters.
}
\examples{
# currently no examples
}
