% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateprobs_p.R
\name{stateprobs_p}
\alias{stateprobs_p}
\title{Calculate conditional local state probabilities for periodically inhomogeneous HMMs}
\usage{
stateprobs_p(delta, Gamma, allprobs, tod, trackID = NULL)
}
\arguments{
\item{delta}{Initial distribution of length N, or matrix of dimension c(k,N) for k independent tracks, if \code{trackID} is provided. This could e.g. be the periodically stationary distribution (for each track).}

\item{Gamma}{Array of transition probability matrices for each time point in the cycle of dimension c(N,N,L), where L is the length of the cycle.}

\item{allprobs}{Matrix of state-dependent probabilities/ density values of dimension c(n, N)}

\item{tod}{(Integer valued) time variable in 1, ..., L, mapping the data index to a generalized time of day (length n).
For half-hourly data L = 48. It could, however, also be day of year for daily data and L = 365.}

\item{trackID}{Optional vector of k track IDs, if multiple tracks need to be decoded separately}
}
\value{
Matrix of conditional state probabilities of dimension c(n,N)
}
\description{
Computes \cr \cr
\eqn{\Pr(S_t = j \mid X_1, ..., X_T)} \cr \cr
}
\examples{
L = 24
beta = matrix(c(-1, 2, -1, -2, 1, -1), nrow = 2, byrow = TRUE)
Gamma = tpm_p(1:L, L, beta, degree = 1)
delta = stationary_p(Gamma, 1)
allprobs = matrix(runif(200), nrow = 100, ncol = 2)
tod = rep(1:24, 5)[1:100]

probs = stateprobs_p(delta, Gamma, allprobs, tod)
}
