% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qreml_functions.R
\name{summary.qremlModel}
\alias{summary.qremlModel}
\title{Summary method for \code{qremlModel} objects}
\usage{
\method{summary}{qremlModel}(object, ...)
}
\arguments{
\item{object}{\code{qremlModel} object created by \code{\link{qreml}}}

\item{...}{additional arguments}
}
\value{
prints a summary of the model object
}
\description{
Prints a summary of a model object created by \code{\link{qreml}}.
}
\examples{
# no examples
}
