% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qreml_functions.R
\name{summary.qremlModel}
\alias{summary.qremlModel}
\title{Summary method for \code{qremlModel} objects}
\usage{
\method{summary}{qremlModel}(object, ...)
}
\arguments{
\item{object}{\code{qremlModel} object}

\item{...}{additional arguments}
}
\value{
prints a summary of the model object
}
\description{
Summary method for \code{qremlModel} objects
}
\examples{
# no examples
}
