% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm.R
\name{tpm}
\alias{tpm}
\title{Build the transition probability matrix from unconstraint parameter vector}
\usage{
tpm(param)
}
\arguments{
\item{param}{Unconstraint parameter vector of length N*(N-1) where N is the number of states of the Markov chain}
}
\value{
Transition probability matrix of dimension c(N,N)
}
\description{
This function builds the transition probability matrix from an unconstraint parameter vector. 
For each row of the matrix, the inverse multinomial logistic link is applied.
}
\examples{
param = rep(-2,6)
Gamma = tpm(param)
}
