% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_functions.R
\name{tpm_emb}
\alias{tpm_emb}
\title{Build the embedded transition probability matrix of an HSMM from unconstrained parameter vector}
\usage{
tpm_emb(param = NULL, byrow = FALSE)
}
\arguments{
\item{param}{Unconstrained parameter vector of length N*(N-2) where N is the number of states of the Markov chain
If the function is called without \code{param}, it will return the conditional transition probability matrix for a 2-state HSMM, which is fixed with 0 diagonal entries and off-diagonal entries equal to 1.}

\item{byrow}{Logical that indicates if the transition probability matrix should be filled by row. 
Defaults to FALSE, but should be set to TRUE if one wants to work with a matrix of beta parameters returned by popular HMM packages like \code{moveHMM}, \code{momentuHMM}, or \code{hmmTMB}.}
}
\value{
Embedded/ conditional transition probability matrix of dimension c(N,N)
}
\description{
Hidden semi-Markov models are defined in terms of state durations and an embedded transition probability matrix that contains the conditional transition probabilities given that the current state is left.
This function builds the embedded/ conditional transition probability matrix from an unconstraint parameter vector. 
For each row of the matrix, the inverse multinomial logistic link is applied.

For a matrix of dimension c(N,N), the number of free off-diagonal elements is N*(N-2), hence also the length of \code{param}.

Compatible with automatic differentiation by \code{RTMB}
}
\examples{
# 2 states: no free off-diagonal elements
omega = tpm_emb()

# 3 states: 3 free off-diagonal elements
param = rep(0, 3)
omega = tpm_emb(param)

# 4 states: 8 free off-diagonal elements
param = rep(0, 8)
omega = tpm_emb(param)
}
