% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi_p.R
\name{viterbi_p}
\alias{viterbi_p}
\title{Viterbi algorithm for decoding states of periodically inhomogeneous HMMs}
\usage{
viterbi_p(delta, Gamma, allprobs, tod, trackID = NULL)
}
\arguments{
\item{delta}{Initial distribution of length N, or matrix of dimension c(k,N) for k independent tracks, if \code{trackID} is provided. This could e.g. be the periodically stationary distribution (for each track).}

\item{Gamma}{Array of transition probability matrices for each time point in the cycle of dimension c(N,N,L), where L is the length of the cycle.}

\item{allprobs}{Matrix of state-dependent probabilities/ density values of dimension c(n, N)}

\item{tod}{Integer valued cyclic variable to index the transition probability matrix.}

\item{trackID}{Optional vector of k track IDs, if multiple tracks need to be decoded separately}
}
\value{
Vector of decoded states of length n
}
\description{
Viterbi algorithm for decoding states of periodically inhomogeneous HMMs
}
\examples{
delta = c(0.5, 0.5)
beta = matrix(c(-2, 1, -1,
                -2, -1, 1), nrow = 2, byrow = TRUE)
Gamma = tpm_p(1:24, 24, beta)

tod = rep(1:24, 10)
n = length(tod)

allprobs = matrix(runif(2*n), nrow = n, ncol = 2)
states = viterbi_p(delta, Gamma, allprobs, tod)
}
